Rename TimeSpanExtensions to UrlTimeCode
